+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Remove legacy API [#888](https://github.com/open-apparel-registry/open-apparel-registry/pull/888)
 
 ### Fixed
+- Handle cases when Source contributor or facility_list are None [#903](https://github.com/open-apparel-registry/open-apparel-registry/pull/903)
 
 ### Security